requirements.txt
pandas>=2.0.0,<2.1.0
numpy>=1.24.0,<2.0.0
gradio>=3.50.0
python-dotenv>=1.0.0
langchain-community==0.0.20
langchain-openai==0.0.8
langchain-text-splitters==0.0.1
langchain==0.0.347
transformers==4.36.2
tqdm>=4.0.0